requirements.txt
asgiref==3.5.2
dj-static==0.0.6
Django==4.0.5
django-stdimage==6.0.1
Pillow==9.1.1
PyMySQL==1.0.2
sqlparse==0.4.2
static3==0.7.0
